Fix perfdata format for OS load (value=current;warning;critical)
This is old Nagios style to return current value with warning and
critical. It should not be used for 1 minute, 5 minute and 15 minutes,
since those are just different measures and should be returned
separately.

Just using space as a field separator didn't fully work. Testing with
icinga2 shows performance data correctly, but when send to influxdb2 all
were considered the same metric.